backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.10.3
python-multipart>=0.0.9
langchain>=1.2.0
langgraph>=0.2.60
langsmith>=0.1.147
chromadb>=0.5.20
langchain-chroma>=0.2.0
langchain-openai>=0.3.0
langchain-ollama>=0.2.0
pypdf>=5.0.1
python-docx>=1.1.2
